usplit up test file
